Changed to float some fields that were declared wrongly as int

+++ damage/models.py
@@ -89,8 +89,8 @@ class HaitiAdmin(models.Model):
     name_5 = models.CharField(max_length=75)
     type_5 = models.CharField(max_length=25)
     engtype_5 = models.CharField(max_length=25)
-    shape_leng = models.IntegerField()
-    shape_area = models.IntegerField()
+    shape_leng = models.FloatField()
+    shape_area = models.FloatField()
     the_geom = models.MultiPolygonField()
     objects = models.GeoManager()
     class Meta:
